feat: implements `ResponseEntity` to `employee` controller

+++ src/main/java/com/carloslonghi/ScrumHub/controller/TaskController.java
@@ -1,10 +1,10 @@
 package com.carloslonghi.ScrumHub.controller;
 
 import com.carloslonghi.ScrumHub.dto.TaskDTO;
-import com.carloslonghi.ScrumHub.mapper.TaskMapper;
-import com.carloslonghi.ScrumHub.model.TaskModel;
 import com.carloslonghi.ScrumHub.service.TaskService;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
 import java.util.List;
@@ -17,27 +17,51 @@ public class TaskController {
     private TaskService taskService;
 
     @GetMapping
-    public List<TaskDTO> getAll() {
-        return taskService.getAllTasks();
+    public ResponseEntity<List<TaskDTO>> getAll() {
+        List<TaskDTO> tasks = taskService.getAllTasks();
+        return ResponseEntity.ok(tasks);
     }
 
     @PostMapping
-    public TaskDTO create(@RequestBody TaskDTO task) {
-        return taskService.createTask(task);
+    public ResponseEntity<String> create(@RequestBody TaskDTO task) {
+        TaskDTO taskCreated = taskService.createTask(task);
+        return ResponseEntity
+                .status(HttpStatus.CREATED)
+                .body("Tarefa \"" + taskCreated.getName() + "\" cadastrada com sucesso");
     }
 
     @GetMapping("/{id}")
-    public TaskDTO getById(@PathVariable Long id) {
-        return taskService.getTaskById(id);
+    public ResponseEntity<Object> getById(@PathVariable Long id) {
+        TaskDTO taskFound = taskService.getTaskById(id);
+        if (taskFound != null) {
+            return ResponseEntity.ok(taskFound);
+        }
+        return ResponseEntity
+                .status(HttpStatus.NOT_FOUND)
+                .body("Não existe tarefa com ID \"" + id + "\"");
     }
 
     @PutMapping("/{id}")
-    public TaskDTO updateById(@PathVariable Long id, @RequestBody TaskDTO task) {
-        return taskService.updateTaskById(id, task);
+    public ResponseEntity<String> updateById(@PathVariable Long id, @RequestBody TaskDTO task) {
+        TaskDTO taskFound = taskService.getTaskById(id);
+        if (taskFound != null) {
+            TaskDTO taskUpdated = taskService.updateTaskById(id, task);
+            return ResponseEntity
+                    .ok("Tarefa \"" + taskUpdated.getName() + "\" atualizada com sucesso");
+        }
+        return ResponseEntity.ok("Não existe tarefa com ID \"" + id + "\"");
     }
 
     @DeleteMapping("/{id}")
-    public void deleteById(@PathVariable Long id) {
-        taskService.deleteTaskById(id);
+    public ResponseEntity<String> deleteById(@PathVariable Long id) {
+        TaskDTO taskFound = taskService.getTaskById(id);
+        if (taskFound != null) {
+            taskService.deleteTaskById(id);
+            return ResponseEntity
+                    .ok("Tarefa \"" + taskFound.getName() + "\" deletada com sucesso");
+        }
+        return ResponseEntity
+                .status(HttpStatus.NOT_FOUND)
+                .body("Não existe tarefa com ID \"" + id + "\"");
     }
 }
